Extract to various functions and file backend logic
